Fix serialization problems when reading SIGNED investments from Zonky
